Common.Data

Description

Version : v1.0.0

This module contains common data structures shared by both server and client program.

Documentation

data User

Represent data structure for a user in the system.

Constructors

User

Fields

nickname :: String

username :: String

profession :: String

location :: String

gender :: Gender

data Gender

Data structure represent the gender and is used in the User data structure.

Constructors

Male

Female

data Status

Data structure Status represent the different status that a user can be in.

Constructors

Offline

NotInTheOffice

Available

OnThePhone

VideoConference

InAMeeting

Busy

DoNotDisurb

mkUserListFromStrList :: [String] -> [User]

Makes a list of users from a list of strings. Strings must be formated the same way you get a string with show User.