Common.Db.FileAccess

Description

Version : v1.0.0

Handles read and write from files in a generic way. Saves files in a subdirectory named db.

Documentation

appendDataToFileLn :: String -> String -> IO ()

Saves string to given file on a new line. If file or subdirectory does not exist, directory and/or file will be created. File is relative to execution directory and will be saved at: db/filename.ext

readFileLnSeparated :: String -> IO [String]

Reads a newline separated file and returns an array of strings where each string is a line in the file. File is relative to execution directory and must exist in subfolder db.

deleteLnFromFile :: String -> String -> IO ()

Deletes given string line from file if it exists. File is relative to execution directory and must exist in subfolder db.