Common.Inet.Data

Description

Version : v1.0.0

Module contains data structures related to communication and is shared between server and client.

Documentation

data InetError

Data structure representing errors in communication between server and client.

Constructors

NoError
NoError is returned when request is processed correctly.

UserParseError
UserParseError is returned when a bad formated user xml is parsed.

UserDoesNotExistError
UserDoesNotExistError is returned when a request for a user that does not exist is sent.

ServiceError
ServiceError is returned when client askes for unknown service.