Common.Inet.Parse

Description

Version : v1.0.0

This module parse String streams sent between server (moses) and client (mos). Streams are built up with simple xml format.

Documentation

userToXml :: User -> String

Converts a User to an xml String. The xml format is simplyfied and looks like: user nicknamenick/nickname usernameuser/username professionengineer/profession locationNew York/location genderMale/gender /user

The returned string is only a single line string without newlines and spaces. All tags continues after each other in one long string.

xmlToMaybeUser :: String -> Maybe User

Converts an xml string to a User. If xml string has wrong format Nothing is returned.

strBetweenTags :: String -> String -> String -> String

Returns the string between starting tag and ending tag. If tag does not exist or string is formated wrong the empty String is returned.