Main

Description

Version : v0.0.1-alpha

This program is the server side of MyOfficeState. Keeps track of the states for users and report state to client program.

Documentation

main :: IO ()

Main function for application moses. Execution of program starts here.

connectLoop :: Socket -> IO ()

Function accepts a connection and handles the connection.

readWriteConn :: (Socket, SockAddr) -> IO ()

Takes care of reading and writing to a connected client.

parseNetService :: String -> InetError

Parse string from client for supported services. If parsing fails ServiceError is returned, otherwise parse will continue. Supported services:

newUser - Client registers a new user at server. User string is supplied. updateUser - Client wants to update its user information. User string is supplied. getUser - Returns the requested user string. User nickname should be supplied as well. status - Returns status for all registered users to client. updateStatus - Client updates its status. Status is supplied.